feat: add auth and swagger docs

+++ src/users/users.service.ts
@@ -1,8 +1,16 @@
-import { Inject, Injectable } from '@nestjs/common'
-import type { Prisma, User } from '@prisma/client'
+import {
+  BadRequestException,
+  Inject,
+  Injectable,
+  UnauthorizedException
+} from '@nestjs/common'
+import { Role, type Prisma, type User } from '@prisma/client'
+
+import { UserPayload } from 'types'
 
 import { PrismaService } from '@/database/prisma.service'
 import { Service } from '@/lib/constants'
+import { UserNotFoundException } from '@/lib/exceptions'
 import { hash } from '@/lib/hash'
 import { CreateUserDto } from './dto/users.dto'
 
@@ -11,6 +19,12 @@ export class UsersService {
   constructor(@Inject(Service.PRISMA) private prisma: PrismaService) {}
 
   async create(data: CreateUserDto): Promise<User> {
+    const user = await this.prisma.user.findUnique({
+      where: { email: data.email }
+    })
+
+    if (user) throw new BadRequestException('Email already exists')
+
     const hashedPassword = await hash(data.password)
 
     return this.prisma.user.create({
@@ -26,12 +40,27 @@ export class UsersService {
   }
 
   async update(
-    options: Pick<Prisma.UserUpdateArgs, 'data' | 'where'>
+    options: Pick<Prisma.UserUpdateArgs, 'data' | 'where'>,
+    currentUser: UserPayload
   ): Promise<User> {
+    const user = await this.prisma.user.findUnique({ where: options.where })
+
+    if (!user) throw new UserNotFoundException()
+
+    if (currentUser.id !== user.id && currentUser.role === Role.USER)
+      throw new UnauthorizedException()
+
     return this.prisma.user.update(options)
   }
 
-  async delete(where: Prisma.UserWhereUniqueInput) {
+  async delete(where: Prisma.UserWhereUniqueInput, currentUser: UserPayload) {
+    const user = await this.prisma.user.findUnique({ where })
+
+    if (!user) throw new UserNotFoundException()
+
+    if (currentUser.id !== user.id && currentUser.role === Role.USER)
+      throw new UnauthorizedException()
+
     return this.prisma.user.delete({ where })
   }
 }
